% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{barplotgene}
\alias{barplotgene}
\title{Gene Expression Barplot}
\usage{
barplotgene(object, g, n = NULL, logsc = FALSE)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{g}{Individual gene name or vector with a group of gene names corresponding to a subset of valid row names of the \code{ndata} slot
of the \code{SCseq} object.}

\item{n}{String of characters representing the title of the plot. Default is \code{NULL} and the first element of \code{g} is chosen.}

\item{logsc}{logical. If \code{TRUE}, then gene expression values are log2-transformed after adding a pseudo-count of 0.1. Default is \code{FALSE}
and untransformed values are shown.}
}
\value{
None
}
\description{
This functions generates a barplot of gene expression across all clusters.
}
