% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{calcAlphaG}
\alias{calcAlphaG}
\title{Function for calculating an aggregated dispersion parameter}
\usage{
calcAlphaG(noise)
}
\arguments{
\item{noise}{List of noise parameters returned by \code{compTBNoise}.}
}
\value{
Matrix of aggregated dispersion parameters.
}
\description{
This function calculates an aggregated dispersion parameter comprising global cell-to-cell variability of transcript counts and biological variability.
}
