% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{calcVar}
\alias{calcVar}
\title{Function for calculating total variance from VarID fit}
\usage{
calcVar(w)
}
\arguments{
\item{w}{List object returned by \code{fitNBtb}.}
}
\value{
Vector of total variance estimates.
}
\description{
This function calculates the total variance from a local VarID fit.
}
