% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{calcVarFit}
\alias{calcVarFit}
\title{Function for calculating the total variance fit}
\usage{
calcVarFit(noise, norm = FALSE)
}
\arguments{
\item{noise}{List of noise parameters returned by \code{compTBNoise}.}

\item{norm}{Logical. If \code{TRUE} then total variance is normalized by the technical noise component (i.e., sampling noise plus global cell-to-cell variability in transcript counts.). Default is FALSE.}
}
\value{
Matrix of total variance fits.
}
\description{
This function calculates a total variance fit comprising sampling noise, global cell-to-cell variability of transcript counts, and biological variability.
}
