% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_genes.R
\docType{data}
\name{cc_genes}
\alias{cc_genes}
\title{Cell cycle markers for Mus Muscuus}
\format{
A list of two components with S phase marker (s) and G2M phase marker (g2m) gene symbols.
}
\usage{
cc_genes
}
\value{
None
}
\description{
This dataset contains official gene symbols for markers of the S phase and G2/M phase of the cell cycle in mouse.
}
\keyword{datasets}
