% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{clustheatmap}
\alias{clustheatmap}
\title{Plotting a Heatmap of the Distance Matrix}
\usage{
clustheatmap(object, final = TRUE, hmethod = "single")
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{final}{logical. If \code{TRUE}, then cells are grouped based on final clusters after outlier identification.
If \code{FALSE}, then initial clusters prior to outlier identification are used for grouping. Default is \code{TRUE}.}

\item{hmethod}{Agglomeration method used for determining the cluster order from hierarchical clustering of the cluster medoids. See \code{hclust} function.}
}
\value{
Returns a vector of cluster numbers ordered as determined by herarchical clustering of cluster the cluster medoids as depicted in the heatmap.
}
\description{
This functions plots a heatmap of the distance matrix grouped by clusters.
}
