% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{corrVar}
\alias{corrVar}
\title{Function for regressing out the mean-variance dependence.
This function corrects for the systematic dependence of the variance on the mean by a local regression.}
\usage{
corrVar(m, v, span = 0.75, degree = 2)
}
\arguments{
\item{m}{Vector of mean expression estimates for a set of genes.}

\item{v}{Vector of variance etsimates for a set of genes.}

\item{span}{Parameter for the local regression. See help(loess). Default value is 0.75.}

\item{degree}{Parameter for the local regression. See help(loess). Default value is 2.}
}
\value{
Vector of corrected variance estimates.
}
\description{
Function for regressing out the mean-variance dependence.
This function corrects for the systematic dependence of the variance on the mean by a local regression.
}
