% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{fitGammaRt}
\alias{fitGammaRt}
\title{Fitting a Gamma distribution to global cell-to-cell variability}
\usage{
fitGammaRt(x)
}
\arguments{
\item{x}{Transcript count matrix with cells as columns and genes as rows.}
}
\value{
Shape parameter of the Gamma distribution. This parameter corresponds to the dispersion explained by the global cell-to-cell variability of UMI counts in a negative binomial model.
}
\description{
This function fits a Gamma distribution to the total transcript counts distribution across a given group of cells. Total transcript counts are normalized by the mean total transcript count across the group. This function is used to infer a Gamma distribution of the global cell-to-cell variability across pruned nearest neighbourhoods.
}
