% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{fitLogVarLogMean}
\alias{fitLogVarLogMean}
\title{Second order polynomial fit of mean-variance dependence
This function corrects for the systematic dependence of the variance on the mean by a local regression.}
\usage{
fitLogVarLogMean(x)
}
\arguments{
\item{x}{Matrix of transcript counts with genes as rows and cells as columns.}
}
\value{
Second order polynomial model as obtained by \code{lm}.
}
\description{
Second order polynomial fit of mean-variance dependence
This function corrects for the systematic dependence of the variance on the mean by a local regression.
}
