% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{getfdata}
\alias{getfdata}
\title{Extracting filtered expression data}
\usage{
getfdata(object, g = NULL, n = NULL)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{g}{Vector of gene names to be included corresponding to a subset of valid row names of the \code{ndata} slot
of the \code{SCseq} object. Default is \code{NULL} and data for all genes remaining after filtering by the \code{filterdata} function are shown.}

\item{n}{Vector of valid column names corresponding to a subset of valid column names of the \code{ndata} slot
of the \code{SCseq} object. Default is \code{NULL} and data for all cells remaining after filtering by the \code{filterdata} function are shown.}
}
\value{
Matrix of filtered expression data with genes as rows and cells as columns.
}
\description{
This functions allows the extraction of a filtered and normalized expression matrix
}
