% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{getFilteredCounts}
\alias{getFilteredCounts}
\title{Function for filtering count data}
\usage{
getFilteredCounts(object, minnumber = 5, minexpr = 5)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{minnumber}{Integer number greater or equal to zero. Minimum number of cells required to have at least \code{minexpr} transcript counts for a gene to not be discarded. Default is 5.}

\item{minexpr}{Integer number greater or equal to zero. Minimum expression of a gene in at least \code{minnumber} cells to not be discarded. Default is 5.}
}
\value{
Filtered expression matrix.
}
\description{
This function discards lowly expressed genes from a count matrix stored in an \code{SCseq} object.
}
