% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{getNode}
\alias{getNode}
\title{Extract all genes for a module in a FateID self-orgaizing map}
\usage{
getNode(ps, n)
}
\arguments{
\item{ps}{FateID SOM. List object.}

\item{n}{Integer number of SOM module.}
}
\value{
Vector of gene IDs in module \code{n}.
}
\description{
Extract a vector of all genes corresponding to a given module of a FateID self-organizing map (SOM) of pseudo-time ordered gene expression (or noise) profiles.
}
