% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{imputeexp}
\alias{imputeexp}
\title{Imputed expression matrix}
\usage{
imputeexp(object, genes = NULL)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{genes}{vector of valid gene names corresponding to row names of slot \code{ndata}. Default is \code{NULL} and imputing is done for all genes.}
}
\value{
An expression matrix with imputed expression values after size normalization. Genes are in rows and cells in columns.
}
\description{
This functions returns an imputed expression matrix based on the imputing computed with \code{compdist}.
}
