% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{plotB}
\alias{plotB}
\title{Boxplots for features across clusters}
\usage{
plotB(x, part, cluster = NULL, set = NULL, ...)
}
\arguments{
\item{x}{Vector of real numbers.}

\item{part}{Vector with cluster partition, e.g., element \code{partition} returned by the \code{graphCluster} function.}

\item{cluster}{Positive integer corresponing to valid cluster number or \code{NULL}. If valid cluster number, then a horizontal line is drawn indicating the median value of \code{x} for the corresponding cluster. If \code{NULL} no line is drawn. Default is \code{NULL}.}

\item{set}{Ordered set of valid cluster numbers. If \code{box} equals \code{TRUE} than data will only be plotted for these clusters in the give}

\item{...}{Additional parameters of \code{boxplot}.}
}
\value{
None
}
\description{
Function to generate boxplots of a feature vector across different clusters.
}
