% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotbackground}
\alias{plotbackground}
\title{Plot Background Model}
\usage{
plotbackground(object)
}
\arguments{
\item{object}{\code{SCseq} class object.}
}
\value{
None
}
\description{
This functions produces a scatter plot showing the gene expression variance as a function of the mean and the inferred
polynomial fit of the background model computed by RaceID3. It also shows a local regression.
}
