% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotdimsat}
\alias{plotdimsat}
\title{Plotting the Saturation of Explained Variance}
\usage{
plotdimsat(object, change = TRUE, lim = NULL)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{change}{logical. If \code{TRUE} then the change in explained variance is plotted. Default is \code{FALSE} and the explained variance is shown.}

\item{lim}{Number of components included for he calculation and shown in the plot. Default is \code{NULL} and all components are included.}
}
\value{
None
}
\description{
This functions plots the explained variance as a function of PCA/ICA components computed by the function \code{CCcorrect}. The number of components where
the change in explained variability upon adding further components approaches linear behaviour demarcates the saturation point and is highlighted in blue.
}
