% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StemID.R
\name{plotdistanceratio}
\alias{plotdistanceratio}
\title{Histogram of Cell-to-Cell Distances in Real versus Embedded Space}
\usage{
plotdistanceratio(object)
}
\arguments{
\item{object}{\code{Ltree} class object.}
}
\value{
None.
}
\description{
This function plots a histogram of the ratios of cell-to-cell distances in the original versus the high-dimensional embedded space used as input for
the StemID2 inferences. The embedded space approximates correlation-based distances by Euclidean distances obtained by classical multi-dimensional scaling.
A minimum spanning tree of the cluster centers is overlaid for comparison.
}
