% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{plotExpNoise}
\alias{plotExpNoise}
\title{Noise-expression scatter plot}
\usage{
plotExpNoise(g, object, noise, set = NULL, ps = 0.1, norm = TRUE, ...)
}
\arguments{
\item{g}{Valid gene ID with available expression and noise estimates.}

\item{object}{\pkg{RaceID} \code{SCseq} object.}

\item{noise}{List object returned by the \code{compTBNoise} function.}

\item{set}{Set of valid cluster numbers. Default is \code{NULL} and data are plotted for cells from all clusters.}

\item{ps}{Real number. Pseudo-count added to noise and expression estimates. Default is 0.1.}

\item{norm}{logical. If \code{FALSE}, then noise is plotted versus non-normalized expression. Default is \code{TRUE} and noise is plotted against normalized expression.}

\item{...}{Additional arguments of \code{plot} function.}
}
\value{
None.
}
\description{
Plotting noise (epsilon) as a function of normalized or non-normalized expression for a given gene.
}
