% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotjaccard}
\alias{plotjaccard}
\title{Plot Jaccard Similarities}
\usage{
plotjaccard(object)
}
\arguments{
\item{object}{\code{SCseq} class object.}
}
\value{
None
}
\description{
This functions plots a barchart of Jaccard similarities for the RaceID3 clusters before outlier identification
}
