% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StemID.R
\name{plotlinkpv}
\alias{plotlinkpv}
\title{Heatmap of Link P-values}
\usage{
plotlinkpv(object)
}
\arguments{
\item{object}{\code{Ltree} class object.}
}
\value{
None.
}
\description{
This function plots a heatmap of link p-values.
}
