% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StemID.R
\name{plotlinkscore}
\alias{plotlinkscore}
\title{Heatmap of Link Scores}
\usage{
plotlinkscore(object)
}
\arguments{
\item{object}{\code{Ltree} class object.}
}
\value{
None.
}
\description{
This function plots a heatmap of link score.
}
