% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotmap}
\alias{plotmap}
\title{Plotting a t-SNE map}
\usage{
plotmap(object, final = TRUE, tp = 1, fr = FALSE, um = FALSE,
  cex = 0.5)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{final}{logical. If \code{TRUE}, then highlight final clusters after outlier identification. If \code{FALSE}, then highlight initial
clusters prior to outlier identification. Default is \code{TRUE}.}

\item{tp}{Number between 0 and 1 to change transparency of dots in the map. Default is 1.}

\item{fr}{logical. If \code{TRUE} then plot Fruchterman-Rheingold layout. Default is \code{FALSE}.}

\item{um}{logical. If \code{TRUE} then plot umap dimensional reduction representation. Default is \code{FALSE}.}

\item{cex}{size of data points. Default value is 0.5.}
}
\value{
None
}
\description{
This functions plots a two-dimensional t-SNE map or a Fruchterman-Rheingold graph layout
of the singe-cell transcriptome data.
}
