% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{plotNoiseModel}
\alias{plotNoiseModel}
\title{Function for plotting the baseline model of gene expression variability}
\usage{
plotNoiseModel(x, corrected = FALSE)
}
\arguments{
\item{x}{List object returned by function \code{noiseBaseFit} or function \code{compNoise}.}

\item{corrected}{logical value. If \code{TRUE}, then the variance is plotted after regressing our the mean dependence.}
}
\value{
None
}
\description{
This function plots the variance against mean expression across all genes and a second order polynomial to the base line of the variance-mean dependence in log space.
}
\examples{
x <- noiseBaseFit(intestinalDataSmall,step=.01,thr=.05)
plotNoiseModel(x)
}
