% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotoutlierprobs}
\alias{plotoutlierprobs}
\title{Plot Outlier Probabilities}
\usage{
plotoutlierprobs(object)
}
\arguments{
\item{object}{\code{SCseq} class object.}
}
\value{
None
}
\description{
This functions plots a barchart of outlier probabilities across all cells in each cluster.
}
