% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{plotPC}
\alias{plotPC}
\title{Function to plot the selected number of principal components}
\usage{
plotPC(res, logDiff = FALSE)
}
\arguments{
\item{res}{List object with k nearest neighbour information returned by \code{pruneKnn} function.}

\item{logDiff}{logical. If \code{TRUE}, then plot log2 of the difference in variability explained by PC i and PC i+1.}
}
\description{
This functions plots the percentage variability explained the first one hundred (or \code{pcaComp}) pricipal components of the PCA performed in the function \code{pruneKnn} if the parameter \code{large} was TRUE. The selected number of principal components (if \code{pcaComp} was NULL) is determined by an elbow criterion and highlighted by a blue circle.
}
