% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StemID.R
\name{plotprojections}
\alias{plotprojections}
\title{Two-Dimensional Map of Cell Projections}
\usage{
plotprojections(object)
}
\arguments{
\item{object}{\code{Ltree} class object.}
}
\value{
None.
}
\description{
This function plots the projections of cells on inter-clustering links connecting cluster medoids in a two-dimensional t-SNE representation.
A minimum spanning tree of the cluster centers is overlaid for comparison.
}
