% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{plotPT}
\alias{plotPT}
\title{Plotting pseudo-time in dimensional reduction representation}
\usage{
plotPT(pt, object, clusters = TRUE, lineages = FALSE)
}
\arguments{
\item{pt}{List object returned by function \code{pseudoTime}.}

\item{object}{\pkg{RaceID} \code{SCseq} object.}

\item{clusters}{logical. If \code{TRUE}, then clusters are highlighted. Otherwise, pseudotime is highlighted. Default is \code{TRUE}.}

\item{lineages}{logical. If \code{TRUE}, then lineages as linear connections of clusters are hghlighted. Otherwise, continuous trajectories are shown. Default is \code{FALSE}.}
}
\value{
None
}
\description{
Highlight clusters or pseudotime in dimensional reduction representation and indicate trajectory derived by \pkg{slingshot}.
}
