% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotsaturation}
\alias{plotsaturation}
\title{Plot Saturation of Within-Cluster Dispersion}
\usage{
plotsaturation(object, disp = FALSE)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{disp}{logical. If \code{FALSE}, then the change of the within-cluster dispersion is plotted. if \code{TRUE} the actual dispersion
is plotted. Default is \code{FALSE}}
}
\value{
None
}
\description{
This functions plots the (change in the) mean within-cluster dispersion as a function of the cluster number
and highlights the saturation point inferred based on the saturation criterion applied by RaceID3: The number of clusters where
the change in within-cluster dispersion upon adding further clusters approaches linear behaviour demarcates the saturation
point and is highlighted in blue.
}
