% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotsensitivity}
\alias{plotsensitivity}
\title{Plot Sensitivity}
\usage{
plotsensitivity(object)
}
\arguments{
\item{object}{\code{SCseq} class object.}
}
\value{
None
}
\description{
This functions plots the number of outliers as a function of the outlier probability.
}
