% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotsilhouette}
\alias{plotsilhouette}
\title{Plot Cluster Silhouette}
\usage{
plotsilhouette(object, final = FALSE)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{final}{logical. If \code{TRUE}, then plot silhouette coefficients for final clusters after outlier identification.
Default is \code{FALSE} and silhouette coefficients are plotted for initial clusters.}
}
\value{
None
}
\description{
This functions produces a silhouette plot for RaceID3 clusters prior or post outlier identification.
}
