% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StemID.R
\name{plotspantree}
\alias{plotspantree}
\title{Minimum Spanning Tree of RaceID3 clusters}
\usage{
plotspantree(object)
}
\arguments{
\item{object}{\code{Ltree} class object.}
}
\value{
None.
}
\description{
This function plots a minimum spanning tree of the RaceID3 cluster medoids in a two-dimensional t-SNE representation.
}
