% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{plotUMINoise}
\alias{plotUMINoise}
\title{Plotting noise dependence on total UMI count}
\usage{
plotUMINoise(object, noise, log.scale = TRUE)
}
\arguments{
\item{object}{\pkg{RaceID} \code{SCseq} object.}

\item{noise}{object returned by \code{compTBNoise} function.}

\item{log.scale}{Logical. If \code{TRUE}  total transcript counts and \code{epsilon} estimates are log2-transformed for plotting. Default is \code{TRUE}.}
}
\description{
This function plots the dependence of mean noise per cell on the total UMI count per cell. It serves as a basis for choosing the prior parameter \code{gamma} (see function \code{compTBNoise}). With a proper parameter choice, there should be no correlation between the two quantities. If a positive correlation is observed, \code{gamma} should be increased in order to weaken the prior. If the correlation is negative, \code{gamma} should be decreased in order to increase the strength of the prior.
}
