% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{postfntb}
\alias{postfntb}
\title{Posterior probability}
\usage{
postfntb(eps, z, x0, gamma, mu, rt)
}
\arguments{
\item{eps}{Positive real number. Residual (biological) noise.}

\item{z}{Vector of integer number greater or equal zero. Transcript counts.}

\item{x0}{Real number. Location parameter.}

\item{gamma}{Positive real number. Scale parameter.}

\item{mu}{Positive real number. Mean expression.}

\item{rt}{Positive real number. Technical noise parameter. See help(fitGammaRt).}
}
\value{
Negative non-normalized log posterior probability fro maximum a posterior inference.
}
\description{
Non-normalized negative log posterior probability with a negative binomial likelihood and Cauchy prior.
}
