% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarID_functions.R
\name{priorfn}
\alias{priorfn}
\title{Prior function for maximum a posterior inference}
\usage{
priorfn(x, x0, gamma)
}
\arguments{
\item{x}{Vector or real numbers (quantiles)}

\item{x0}{Real number. Location parameter.}

\item{gamma}{Positive real number. Scale parameter.}
}
\value{
Vector of probabilities
}
\description{
A prior function specified as Cauchy probability density.
}
