% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StemID.R
\name{projenrichment}
\alias{projenrichment}
\title{Enrichment of cells on inter-cluster links}
\usage{
projenrichment(object)
}
\arguments{
\item{object}{\code{Ltree} class object.}
}
\value{
None.
}
\description{
This function plots a heatmap of the enrichment ratios of cells on significant links.
}
