% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceID.R
\name{plotlabelsmap}
\alias{plotlabelsmap}
\title{Plot labels in the t-SNE map}
\usage{
plotlabelsmap(object, labels = NULL, fr = FALSE, um = FALSE, cex = 0.5)
}
\arguments{
\item{object}{\code{SCseq} class object.}

\item{labels}{Vector of labels for all cells to be highlighted in the t-SNE map. The order has to be the same as for the
columns in slot \code{ndata} of the \code{SCseq} object. Default is \code{NULL} and cell names are highlighted.}

\item{fr}{logical. If \code{TRUE} then plot Fruchterman-Rheingold layout. Default is \code{FALSE}.}

\item{um}{logical. If \code{TRUE} then plot umap dimensional reduction representation. Default is \code{FALSE}.}

\item{cex}{positive real number. Size of the labels. Default is 0.5.}
}
\value{
None
}
\description{
This functions plots cell labels into a two-dimensional t-SNE map or a Fruchterman-Rheingold graph layout
of the singe-cell transcriptome data.
}
